Add input-flag
Add --input-flag for defining the input as a keyword. Also fix a bug
related to python keyword parsing and add chain() and yield_from()
-functions to query scope. Also interpret (**x) as keyword-unpacking
to support better string formatting. Fix a bug with raw-outputting
dicts. Fix a bug with list of lists. Handle query syntax error
exceptions and convert them to user-friendly messages.